+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Components/Interfaces/IMaterialConversionComponent.cs
@@ -1,0 +1,9 @@
+namespace KRT.VRCQuestTools.Components
+{
+    /// <summary>
+    /// Interface for material conversion component.
+    /// </summary>
+    public interface IMaterialConversionComponent
+    {
+    }
+}